pyproject.toml
[project]
name = "fastapi-solid"
version = "1.1.1"
description = "Clean FastAPI template: SOLID design, SQLAlchemy UoW, Redis cache, Alembic."
readme = "README.md"
authors = [
    { name = "knyaz7", email = "hidden@mail.ru" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "dependency-injector>=4.48.2",
    "fastapi>=0.116.2",
    "fastapi-cache2[redis]>=0.2.2",
    "pydantic-settings>=2.11.0",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "ruff>=0.14.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[project.scripts]
fastapi-solid = "fastapi_solid.main:main"

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
src = ["src"]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

skip-magic-trailing-comma = true

[tool.ruff.lint.isort]
split-on-trailing-comma = false
